Refactor: Split AuxController into multiple files
Refactored the AuxController class by splitting it into multiple files, following the "one class, two files" principle. This improves modularity and maintainability by separating the different components into their own header and source files.

The following classes and components were moved to their own files:
- ICVAccess (interface)
- PhysicalOutput
- Effect and its subclasses
- LogicalFunction
- FunctionMapping structs and enums
